remove invalid assertion
The insert should fail becasue of the `turns_player_id_policy` policy
but it is reported as failing because of the `turns_in_turn_policy`
policy. It's either incorrect or I'm doing something wrong.
Anyhow the rwo cannot be inserted so that's good, although it would be
better to ensure the `turns_player_id_policy` policy work as intended...